Add amalgamation support (single-header/source)
This PR introduces an official amalgamation script and integration for FTXUI.

- **tools/amalgamate.py**: A fortified script that generates:
  - ftxui.hpp / ftxui.cpp: Split amalgamated version.
  - ftxui_all.hpp: Truly single-file header-only version.
- **Traceability**: All generated files include the project version and git commit hash.
- **Robustness**: Stricter regex, surgical license stripping, and visibility macro handling.
- **CI/CD Integration**: Automated verification across Linux/MacOS/Windows and packaging in release zip.
- **Documentation**: New dedicated installation guide and examples.

Fixed:https://github.com/ArthurSonzogni/FTXUI/issues/1252

+++ src/ftxui/screen/color.cpp
@@ -258,11 +258,11 @@ Color Color::Interpolate(float t, const Color& a, const Color& b) {
   // https://en.wikipedia.org/wiki/Gamma_correction
   auto interp = [t](uint8_t a_u, uint8_t b_u) {
     constexpr float gamma = 2.2F;
-    const float a_f = powf(a_u, gamma);
-    const float b_f = powf(b_u, gamma);
+    const float a_f = std::pow(a_u, gamma);
+    const float b_f = std::pow(b_u, gamma);
     const float c_f = a_f * (1.0F - t) +  //
                       b_f * t;
-    return static_cast<uint8_t>(powf(c_f, 1.F / gamma));
+    return static_cast<uint8_t>(std::pow(c_f, 1.F / gamma));
   };
   return Color::RGB(interp(a_r, b_r),   //
                     interp(a_g, b_g),   //
